requirements.txt
asyncclick>=8.1.8
click>=8.1.8
fastapi>=0.115.12
google-adk>=1.0.0
google-genai>=1.11.0
httpx>=0.28.1
httpx-sse>=0.4.0
pydantic>=2.11.4
python-dotenv>=1.1.0
starlette>=0.46.2
uvicorn>=0.34.2
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
supabase>=2.0.0